Move deserializers below the io package
